+++ mail/passwordReset-text.php
@@ -3,7 +3,7 @@
 /* @var $this yii\web\View */
 /* @var $user \wdmg\users\models\Users */
 
-$resetLink = Yii::$app->urlManager->createAbsoluteUrl([Yii::$app->requestedRoute, 'token' => $user->password_reset_token]);
+$resetLink = Yii::$app->urlManager->createAbsoluteUrl([$linkRoute, 'token' => $user->password_reset_token]);
 
 ?>
 <?= Yii::t('app/modules/users', 'Hi {username}!', [