+++ messages/ru-RU/app/modules/users.php
@@ -20,6 +20,10 @@ return [
     'Disabled' => 'Не активный',
     'Deleted' => 'Удален',
 
+    'Roles' => 'Роли доступа',
+    'Permissions' => 'Разрешения доступа',
+    'Assignments' => 'Назначения доступа',
+
     'Add new user' => 'Добавить пользователя',
     'Users search' => 'Поиск пользователей',
 
@@ -54,6 +58,17 @@ return [
     'Password reset for {appname}' => 'Сброс пароля к сайту {appname}',
     'Hi {username}!' => 'Здравствуйте {username}!',
     'Follow the link below to reset your password: {link}' => 'Перейдите по данной ссылке для сброса вашего пароля: {link}',
+
+    'Check your email for confirmation registration.' => 'Проверьте свою электронную почту для подтверждения регистрации.',
+    'Error sending registration confirmation email.' => 'Ошибка отправки электронного письма с подтверждением регистрации.',
+    'Error! An empty registration confirmation token.' => 'Ошибка! Пустой токен подтверждения регистрации.',
+    'Error! User not found.' => 'Ошибка! Пользователь не найден.',
+    'Error updating user data.' => 'Ошибка обновления пользовательских данных.',
+    'User authentication error.' => 'Ошибка аутентификации пользователя.',
+    'User registration failed!' => 'Регистрация пользователя не удалась!',
+
+    'Registration is not complete. Confirm the email address of the link from the letter.' => 'Регистрация не завершена. Подтвердите адрес эл.почты по ссылке из письма.',
+    'The user has been blocked by the administrator.' => 'Пользователь был заблокирован администратором.',
 
     'Some fields may contain limited information. We recommend installing the {modules} {count, plural, =1{module} one{module} few{modules} many{modules} other{modules}} for complete compatibility.' => 'Некоторые поля могут содержат ограниченные сведения. Рекомендуем установить {count, plural, =1{модуль} one{модуль} few{модули} many{модули} other{модули}} {modules} для полной совместимости.',
     '{y, plural, =0{} =1{# year} one{# year} few{# years} many{# years} other{# years}}{y, plural, =0{} =1{, } other{, }}{m, plural, =0{} =1{# month} one{# month} few{# months} many{# months} other{# months}}{m, plural, =0{} =1{, } other{, }}{d, plural, =0{} =1{# day} one{# day} few{# days} many{# days} other{# days}}{d, plural, =0{} =1{, } other{, }}{h, plural, =0{} =1{# hour} one{# hour} few{# hours} many{# hours} other{# hours}}{h, plural, =0{} =1{, } other{, }}{i, plural, =0{} =1{# minute} one{# minute} few{# minutes} many{# minutes} other{# minutes}}{i, plural, =0{} =1{, } other{, }}{s, plural, =0{} =1{# second} one{# second} few{# seconds} many{# seconds} other{# seconds}}{invert, plural, =0{ left} =1{ ago} other{}}' => '{y, plural, =0{} =1{# год} one{# год} few{# года} many{# лет} other{# лет}}{y, plural, =0{} =1{, } other{, }}{m, plural, =0{} =1{# месяц} one{# месяц} few{# месяца} many{# месяцев} other{# месяцев}}{m, plural, =0{} =1{, } other{, }}{d, plural, =0{} =1{# день} one{# день} few{# дня} many{# дней} other{# дней}}{d, plural, =0{} =1{, } other{, }}{h, plural, =0{} =1{# час} one{# час} few{# часа} many{# часов} other{# часов}}{h, plural, =0{} =1{, } other{, }}{i, plural, =0{} =1{# минута} one{# минута} few{# минуты} many{# минут} other{# минут}}{i, plural, =0{} =1{, } other{, }}{s, plural, =0{} =1{# секунда} one{# секунд} few{# секунды} many{# секунд} other{# секунд}}{invert, plural, =0{ осталось} =1{ назад} other{}}',